Move checksummed encoding from Core to Encoding
