+++ src/Core.php
@@ -433,110 +433,4 @@ final class Core
         }
     }
 
-    /**
-     * Save bytes to check summed ascii safe string.
-     *
-     * @param $header
-     * @param $bytes
-     *
-     * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
-     *
-     * @return string
-     */
-    public static function saveBytesToChecksummedAsciiSafeString($header, $bytes)
-    {
-        // Headers must be a constant length to prevent one type's header from
-        // being a prefix of another type's header, leading to ambiguity.
-        if (Core::ourStrlen($header) !== Core::SERIALIZE_HEADER_BYTES) {
-            throw new Ex\EnvironmentIsBrokenException(
-                'Header must be 4 bytes.'
-            );
-        }
-
-        return Encoding::binToHex(
-            $header .
-            $bytes .
-            \hash(
-                Core::CHECKSUM_HASH_ALGO,
-                $header . $bytes,
-                true
-            )
-        );
-    }
-
-    /**
-     * Load bytes from checksummed ascii safe string.
-     *
-     * @param $expected_header
-     * @param $string
-     *
-     * @throws \Defuse\Crypto\Exception\EnvironmentIsBrokenException
-     * @throws \Defuse\Crypto\Exception\BadFormatException
-     *
-     * @return string
-     */
-    public static function loadBytesFromChecksummedAsciiSafeString($expected_header, $string)
-    {
-        // Headers must be a constant length to prevent one type's header from
-        // being a prefix of another type's header, leading to ambiguity.
-        if (Core::ourStrlen($expected_header) !== Core::SERIALIZE_HEADER_BYTES) {
-            throw new Ex\EnvironmentIsBrokenException(
-                'Header must be 4 bytes.'
-            );
-        }
-
-        try {
-            $bytes = Encoding::hexToBin($string);
-        } catch (\RangeException $ex) {
-            throw new Ex\BadFormatException(
-                'String has invalid hex encoding.'
-            );
-        }
-
-        /* Make sure we have enough bytes to get the version header and checksum. */
-        if (Core::ourStrlen($bytes) < Core::SERIALIZE_HEADER_BYTES + Core::CHECKSUM_BYTE_SIZE) {
-            throw new Ex\BadFormatException(
-                'Encoded data is shorter than expected.'
-            );
-        }
-
-        /* Grab the version header. */
-        $actual_header = Core::ourSubstr($bytes, 0, Core::SERIALIZE_HEADER_BYTES);
-
-        if ($actual_header !== $expected_header) {
-            throw new Ex\BadFormatException(
-                'Invalid header.'
-            );
-        }
-
-        /* Grab the bytes that are part of the checksum. */
-        $checked_bytes = Core::ourSubstr(
-            $bytes,
-            0,
-            Core::ourStrlen($bytes) - Core::CHECKSUM_BYTE_SIZE
-        );
-
-        /* Grab the included checksum. */
-        $checksum_a = Core::ourSubstr(
-            $bytes,
-            Core::ourStrlen($bytes) - Core::CHECKSUM_BYTE_SIZE,
-            Core::CHECKSUM_BYTE_SIZE
-        );
-
-        /* Re-compute the checksum. */
-        $checksum_b = \hash(Core::CHECKSUM_HASH_ALGO, $checked_bytes, true);
-
-        /* Validate it. It *is* important for this to be constant time. */
-        if (! Core::hashEquals($checksum_a, $checksum_b)) {
-            throw new Ex\BadFormatException(
-                "Data is corrupted, the checksum doesn't match"
-            );
-        }
-
-        return Core::ourSubstr(
-            $bytes,
-            Core::SERIALIZE_HEADER_BYTES,
-            Core::ourStrlen($bytes) - Core::SERIALIZE_HEADER_BYTES - Core::CHECKSUM_BYTE_SIZE
-        );
-    }
 }
